Rename KeyboardSpacerPropTypes -> BaseKeyboardSpacerPropTypes
